refactor(hogql): streamline function call handling in printer classes
