+++ posthog/hogql/printer/test/test_printer.py
@@ -2573,7 +2573,7 @@ class TestPrinter(BaseTest):
 
     def test_get_survey_response(self):
         # Test with just question index
-        with patch("posthog.hogql.printer.base.get_survey_response_clickhouse_query") as mock_get_survey_response:
+        with patch("posthog.hogql.printer.clickhouse.get_survey_response_clickhouse_query") as mock_get_survey_response:
             mock_get_survey_response.return_value = "MOCKED SQL FOR SURVEY RESPONSE"
 
             printed = self._print(
@@ -2588,7 +2588,7 @@ class TestPrinter(BaseTest):
             self.assertIn("MOCKED SQL FOR SURVEY RESPONSE", printed)
 
         # Test with question index and specific ID
-        with patch("posthog.hogql.printer.base.get_survey_response_clickhouse_query") as mock_get_survey_response:
+        with patch("posthog.hogql.printer.clickhouse.get_survey_response_clickhouse_query") as mock_get_survey_response:
             mock_get_survey_response.return_value = "MOCKED SQL FOR SURVEY RESPONSE WITH ID"
 
             printed = self._print(
@@ -2603,7 +2603,7 @@ class TestPrinter(BaseTest):
             self.assertIn("MOCKED SQL FOR SURVEY RESPONSE WITH ID", printed)
 
         # Test with multiple choice question
-        with patch("posthog.hogql.printer.base.get_survey_response_clickhouse_query") as mock_get_survey_response:
+        with patch("posthog.hogql.printer.clickhouse.get_survey_response_clickhouse_query") as mock_get_survey_response:
             mock_get_survey_response.return_value = "MOCKED SQL FOR MULTIPLE CHOICE SURVEY RESPONSE"
 
             printed = self._print(
@@ -2616,7 +2616,7 @@ class TestPrinter(BaseTest):
 
     def test_unique_survey_submissions_filter(self):
         with patch(
-            "posthog.hogql.printer.base.filter_survey_sent_events_by_unique_submission"
+            "posthog.hogql.printer.clickhouse.filter_survey_sent_events_by_unique_submission"
         ) as mock_filter_survey_sent_events_by_unique_submission:
             mock_filter_survey_sent_events_by_unique_submission.return_value = (
                 "MOCKED SQL FOR UNIQUE SURVEY SUBMISSIONS FILTER"
